Fix 16-bit->8-bit color truncation in Voronoi and SphericalHarmonics
Both effects cast uint16_t linear color channels (0-65535) through
static_cast<uint8_t>, wrapping values mod-256 into mostly-dark garbage:
- Voronoi cell-color smoothing ran the truncation on nearly every pixel
  (~200 sites -> secondBest always valid).
- SphericalHarmonics truncated the negative-lobe green channel.

Widen both casts to uint16_t. The lerp/scale stays within [0, 65535], so
the change is behavior-preserving apart from removing the corruption.

Verified: native suite passes (both render in the 288x144 smoke harness).

Records P1 #2 fixed in docs/CODE_REVIEW.md.

+++ effects/SphericalHarmonics.h
@@ -164,7 +164,7 @@ public:
           baked_palette.get(std::min(1.0f, abs_val * params.amplitude));
       // Negative palette (channel-swapped)
       Color4 neg =
-          Color4(Pixel(pos.color.b, static_cast<uint8_t>(pos.color.g * 0.8f),
+          Color4(Pixel(pos.color.b, static_cast<uint16_t>(pos.color.g * 0.8f),
                        pos.color.r),
                  pos.alpha);
